Add missing cast to make MSVC happier. Review URL: http://codereview.chromium.org/10592003
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@11872 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/isolate.cc
@@ -1608,7 +1608,7 @@ void Isolate::PushToPartialSnapshotCache(Object* obj) {
   int capacity = serialize_partial_snapshot_cache_capacity();
 
   if (length >= capacity) {
-    int new_capacity = (capacity + 10) * 1.2;
+    int new_capacity = static_cast<int>((capacity + 10) * 1.2);
     Object** new_array = new Object*[new_capacity];
     for (int i = 0; i < length; i++) {
       new_array[i] = serialize_partial_snapshot_cache()[i];
